fix: consolidate VersionUtil, update imports, and add migration warning
