+++ app/src/main/java/io/apicurio/registry/storage/impl/sql/repositories/SqlVersionRepository.java
@@ -45,7 +45,7 @@ import io.apicurio.registry.storage.impl.sql.mappers.StringMapper;
 import io.apicurio.registry.storage.impl.sql.mappers.VersionContentDtoMapper;
 import io.apicurio.registry.storage.impl.sql.mappers.VersionStateMapper;
 import io.apicurio.registry.types.VersionState;
-import io.apicurio.registry.util.VersionUtil;
+import io.apicurio.registry.utils.VersionUtil;
 import static io.apicurio.registry.utils.StringUtil.asLowerCase;
 import static io.apicurio.registry.utils.StringUtil.limitStr;
 import io.apicurio.registry.utils.impexp.v3.ArtifactVersionEntity;
